backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-multipart==0.0.20
python-dotenv==1.1.1
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
PyJWT==2.10.1
pinecone==5.3.1
supabase==2.15.0
pypdf==5.9.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.104.0
